wip: extract general project init
